fix: Improve empty state and rename stacks-orders to camelCase
